+++ BraintreeApplePay/BTApplePayCardNonce.m
@@ -2,8 +2,11 @@
 
 @implementation BTApplePayCardNonce
 
-- (instancetype)initWithNonce:(NSString *)nonce localizedDescription:(NSString *)description type:(NSString *)type json:(BTJSON *)json {
-    self = [super initWithNonce:nonce localizedDescription:description type:type isDefault:[json[@"default"] isTrue]];
+- (instancetype)initWithNonce:(NSString *)nonce
+                         type:(NSString *)type
+                         json:(BTJSON *)json {
+    self = [super initWithNonce:nonce type:type isDefault:[json[@"default"] isTrue]];
+    
     if (self) {
         _binData = [[BTBinData alloc] initWithJSON:json[@"binData"]];
     }