+++ BraintreeApplePayTests/BTPaymentMethodNonceParser_ApplePay_Tests.swift
@@ -5,7 +5,6 @@ class BTPaymentMethodNonceParser_ApplePay_Tests: XCTestCase {
         let sharedParser = BTPaymentMethodNonceParser.shared()
         let applePayCard = BTJSON(value: [
             "consumed": false,
-            "description": "Apple Pay Card ending in 11",
             "details": [
                 "cardType": "American Express"
             ],
@@ -19,6 +18,5 @@ class BTPaymentMethodNonceParser_ApplePay_Tests: XCTestCase {
 
         XCTAssertEqual(applePayCardNonce?.nonce, "a-nonce")
         XCTAssertEqual(applePayCardNonce?.type, "American Express")
-        XCTAssertEqual(applePayCardNonce?.localizedDescription, "Apple Pay Card ending in 11")
     }
 }