+++ BraintreeCard/BTCardNonce_Internal.h
@@ -6,7 +6,6 @@ NS_ASSUME_NONNULL_BEGIN
 @interface BTCardNonce ()
 
 - (instancetype)initWithNonce:(nonnull NSString *)nonce
-                  description:(nullable NSString *)description
                   cardNetwork:(BTCardNetwork)cardNetwork
               expirationMonth:(nullable NSString *)expirationMonth
                expirationYear:(nullable NSString *)expirationYear