+++ BraintreeCore/BTPaymentMethodNonce.m
@@ -3,7 +3,6 @@
 @interface BTPaymentMethodNonce ()
 
 @property (nonatomic, copy, readwrite) NSString *nonce;
-@property (nonatomic, copy, readwrite) NSString *localizedDescription;
 @property (nonatomic, copy, readwrite) NSString *type;
 @property (nonatomic, readwrite, assign) BOOL isDefault;
 
@@ -11,23 +10,24 @@
 
 @implementation BTPaymentMethodNonce
 
-- (instancetype)initWithNonce:(NSString *)nonce localizedDescription:(NSString *)description type:(NSString *)type {
+- (instancetype)initWithNonce:(NSString *)nonce type:(NSString *)type {
     if (!nonce) return nil;
     
     if (self = [super init]) {
         self.nonce = nonce;
-        self.localizedDescription = description;
         self.type = type;
     }
     return self;
 }
 
-- (nullable instancetype)initWithNonce:(NSString *)nonce localizedDescription:(nullable NSString *)description {
-    return [self initWithNonce:nonce localizedDescription:description type:@"Unknown"];
+- (nullable instancetype)initWithNonce:(NSString *)nonce {
+    return [self initWithNonce:nonce type:@"Unknown"];
 }
 
-- (nullable instancetype)initWithNonce:(NSString *)nonce localizedDescription:(NSString *)description type:(nonnull NSString *)type isDefault:(BOOL)isDefault {
-    if (self = [self initWithNonce:nonce localizedDescription:description type:type]) {
+- (nullable instancetype)initWithNonce:(NSString *)nonce
+                                  type:(nonnull NSString *)type
+                             isDefault:(BOOL)isDefault {
+    if (self = [self initWithNonce:nonce type:type]) {
         _isDefault = isDefault;
     }
     return self;