+++ BraintreeCore/Public/BTPaymentMethodNonce.h
@@ -19,41 +19,33 @@ NS_ASSUME_NONNULL_BEGIN
  Initialize a new Payment Method Nonce.
 
  @param nonce       A transactable payment method nonce.
- @param description A human-readable description.
- @param type        A string identifying the type of the payment method.
  @return A Payment Method Nonce, or `nil` if nonce is nil.
 */
-- (nullable instancetype)initWithNonce:(NSString *)nonce localizedDescription:(nullable NSString *)description type:(NSString *)type;
+- (nullable instancetype)initWithNonce:(NSString *)nonce;
 
 /**
  Initialize a new Payment Method Nonce.
 
  @param nonce       A transactable payment method nonce.
- @param description A human-readable description.
+ @param type         A string identifying the type of the payment method.
  @return A Payment Method Nonce, or `nil` if nonce is nil.
 */
-- (nullable instancetype)initWithNonce:(NSString *)nonce localizedDescription:(nullable NSString *)description;
+- (nullable instancetype)initWithNonce:(NSString *)nonce type:(NSString *)type;
 
 /**
  Initialize a new Payment Method Nonce.
 
  @param nonce       A transactable payment method nonce.
- @param description A human-readable description.
  @param type        A string identifying the type of the payment method.
  @param isDefault   A boolean indicating whether this is a default payment method.
  @return A Payment Method Nonce, or `nil` if nonce is nil.
 */
-- (nullable instancetype)initWithNonce:(NSString *)nonce localizedDescription:(NSString *)description type:(nonnull NSString *)type isDefault:(BOOL)isDefault;
+- (nullable instancetype)initWithNonce:(NSString *)nonce type:(nonnull NSString *)type isDefault:(BOOL)isDefault;
 
 /**
  The one-time use payment method nonce
 */
 @property (nonatomic, readonly, copy) NSString *nonce;
-
-/**
- A localized description of the payment info
-*/
-@property (nonatomic, readonly, copy) NSString *localizedDescription;
 
 /**
  The type of the tokenized data, e.g. PayPal, Venmo, MasterCard, Visa, Amex