+++ BraintreeThreeDSecureTests/BTThreeDSecure_Tests.swift
@@ -569,4 +569,3 @@ class BTThreeDSecure_UnitTests: XCTestCase {
     }
 }
 
-