+++ BraintreeVenmo/BTVenmoAccountNonce.m
@@ -9,11 +9,10 @@
 @implementation BTVenmoAccountNonce
 
 - (instancetype)initWithPaymentMethodNonce:(NSString *)nonce
-                               description:(__unused NSString *)description
                                   username:(NSString *)username
                                  isDefault:(BOOL)isDefault
 {
-    if (self = [super initWithNonce:nonce localizedDescription:username type:@"Venmo" isDefault:isDefault]) {
+    if (self = [super initWithNonce:nonce type:@"Venmo" isDefault:isDefault]) {
         _username = username;
     }
     return self;
@@ -21,7 +20,6 @@
 
 + (instancetype)venmoAccountWithJSON:(BTJSON *)venmoAccountJSON {
     return [[[self class] alloc] initWithPaymentMethodNonce:[venmoAccountJSON[@"nonce"] asString]
-                                                description:[venmoAccountJSON[@"description"] asString]
                                                    username:[venmoAccountJSON[@"details"][@"username"] asString]
                                                   isDefault:[venmoAccountJSON[@"default"] isTrue]];
 }